Fixed byte writing.

+++ src/main/java/ru/mpei/uartlab/model/Command.java
@@ -1,4 +1,4 @@
 package ru.mpei.uartlab.model;
 
-public record Command(String tag, int value) {
+public record Command(String tag, short value) {
 }
